+++ src/main/java/com/exasol/projectkeeper/AbstractProjectKeeperMojo.java
@@ -10,7 +10,7 @@ import org.apache.maven.plugins.annotations.Parameter;
 import org.apache.maven.project.MavenProject;
 
 import com.exasol.projectkeeper.validators.DeletedFilesValidator;
-import com.exasol.projectkeeper.validators.changelog.ChangelogValidator;
+import com.exasol.projectkeeper.validators.changesfile.ChangesFileValidator;
 import com.exasol.projectkeeper.validators.files.ProjectFilesValidator;
 import com.exasol.projectkeeper.validators.pom.PomFileIO;
 import com.exasol.projectkeeper.validators.pom.PomFileValidator;
@@ -53,7 +53,9 @@ public abstract class AbstractProjectKeeperMojo extends AbstractMojo {
         final Set<ProjectKeeperModule> enabledModules = getEnabledModules();
         final ExcludedFilesMatcher excludedFilesMatcher = new ExcludedFilesMatcher(this.excludedFiles);
         return List.of(new ProjectFilesValidator(enabledModules, this.project.getBasedir(), excludedFilesMatcher),
-                new PomFileValidator(enabledModules, pomFile), new ChangelogValidator(this.project),
+                new PomFileValidator(enabledModules, pomFile),
+                new ChangesFileValidator(this.project.getVersion(), this.project.getName(),
+                        this.project.getBasedir().toPath()),
                 new DeletedFilesValidator(this.project.getBasedir().toPath(), excludedFilesMatcher));
     }
 }
